+++ CHANGES.rst
@@ -255,6 +255,11 @@ astropy.coordinates
   ensure that any angle was presented in degrees in sky coordinates and
   frames. This is more logically done in the frame itself. [#6858]
 
+- Coordinate frame classes now can't be added to the frame transform graph if
+  they have frame attribute names that conflict with any component names. This
+  is so ``SkyCoord`` can uniquely identify and distinguish frame attributes from
+  frame components. [#6871]
+
 - Slicing and reshaping of ``SkyCoord`` and coordinate frames no longer passes
   the new object through ``__init__``, but directly sets atttributes on a new
   instance. This speeds up those methods by an order of magnitude, but means
@@ -374,6 +379,10 @@ astropy.coordinates
 - Frame objects now use the default differential even if the representation is
   explicitly provided as long as the representation provided is the same type as
   the default representation. [#6944]
+
+- Coordinate frame classes now raise an error when they are added to the frame
+  transform graph if they have frame attribute names that conflict with any
+  component names. [#6871]
 
 astropy.cosmology
 ^^^^^^^^^^^^^^^^^